+++ Project_CinQ/src/main/webapp/WEB-INF/views/default/adminMenu.jsp
@@ -19,7 +19,7 @@
 		<ul>
 			<li><a href="/root/admin/memberAllList">회원관리</a></li>
 			<li><a href="/root/admin/placeAllList">장소관리</a></li>
-			<li><a href="/root/admin/reservationAllListNum">예매관리</a></li>
+			<li><a href="/root/reservation/reservationAllListNum">예매관리</a></li>
 		</ul>
 	</div>
 </body>
